+++ presentation/src/test/java/com/anytypeio/anytype/presentation/editor/editor/EditorFeaturedRelationsTest.kt
@@ -100,8 +100,8 @@ class EditorFeaturedRelationsTest : EditorPresentationTestSetup() {
 
         val objectTypeFields = Block.Fields(
             mapOf(
-                Block.Fields.NAME_KEY to objectTypeName,
-                Block.Fields.DESCRIPTION_KEY to objectTypeDescription
+                Relations.NAME to objectTypeName,
+                Relations.DESCRIPTION to objectTypeDescription
             )
         )
         val customDetails = Block.Details(
@@ -210,8 +210,8 @@ class EditorFeaturedRelationsTest : EditorPresentationTestSetup() {
 
         val objectTypeFields = Block.Fields(
             mapOf(
-                Block.Fields.NAME_KEY to objectTypeName,
-                Block.Fields.DESCRIPTION_KEY to objectTypeDescription
+                Relations.NAME to objectTypeName,
+                Relations.DESCRIPTION to objectTypeDescription
             )
         )
         val customDetails = Block.Details(
@@ -308,8 +308,8 @@ class EditorFeaturedRelationsTest : EditorPresentationTestSetup() {
 
         val objectTypeFields = Block.Fields(
             mapOf(
-                Block.Fields.NAME_KEY to objectTypeName,
-                Block.Fields.DESCRIPTION_KEY to objectTypeDescription
+                Relations.NAME to objectTypeName,
+                Relations.DESCRIPTION to objectTypeDescription
             )
         )
         val customDetails = Block.Details(
@@ -406,8 +406,8 @@ class EditorFeaturedRelationsTest : EditorPresentationTestSetup() {
 
         val objectTypeFields = Block.Fields(
             mapOf(
-                Block.Fields.NAME_KEY to objectTypeName,
-                Block.Fields.DESCRIPTION_KEY to objectTypeDescription
+                Relations.NAME to objectTypeName,
+                Relations.DESCRIPTION to objectTypeDescription
             )
         )
         val customDetails = Block.Details(
@@ -517,8 +517,8 @@ class EditorFeaturedRelationsTest : EditorPresentationTestSetup() {
 
         val objectTypeFields = Block.Fields(
             mapOf(
-                Block.Fields.NAME_KEY to objectTypeName,
-                Block.Fields.DESCRIPTION_KEY to objectTypeDescription
+                Relations.NAME to objectTypeName,
+                Relations.DESCRIPTION to objectTypeDescription
             )
         )
         val customDetails = Block.Details(